fix : 관리자 페이지 글쓰기 수정 feat: notice 이벤트글과 연동 (이벤트글 추가시 최신순으로 바뀌게) ++ 추후 이벤트 페이지 추가시 자세히보기로 연동예정

+++ frontend/src/admin/components/EventNews/EventList.js
@@ -33,12 +33,13 @@ function EventList(props) {
       return "진행 중";
     }
   };
-
   const handleDelete = async (postId) => {
     try {
-      const response = await axios.delete(`/admin/post/${postId}`);
-      console.log(response.data); // 성공적으로 삭제된 경우 메시지 출력
-      // 삭제 후 필요한 동작 수행
+      if (window.confirm("삭제하시겠습니까?")) {
+        const res = await axios.delete(`/admin/post/${postId}`);
+        console.log(res);
+        window.location.reload();
+      }
     } catch (error) {
       console.error(error);
     }
